wip: error handling

+++ src/main/java/com/epf/api/controller/MapsController.java
@@ -5,7 +5,8 @@ import java.util.List;
 import java.util.Map;
 
 import com.epf.api.dto.MapsDTO;
-import com.epf.core.exception.BadAttributException;
+import com.epf.api.exception.BadPathVariableException;
+import com.epf.core.exception.BadAttributeException;
 import com.epf.api.mapper.MapsDTOMapper;
 import com.epf.core.model.Maps;
 import com.epf.core.service.MapsService;
@@ -45,7 +46,7 @@ public class MapsController {
             System.out.println(e);
             Map<String, Object> response = new HashMap<>();
             response.put("status", HttpStatus.NOT_FOUND.value());
-            response.put("message", "not found");
+            response.put("message", "not map found");
             return new ResponseEntity<>(response, HttpStatus.NOT_FOUND);
         } catch (Exception e) {
             System.out.println(e);
@@ -69,8 +70,9 @@ public class MapsController {
             int id = service.create(model);
             Map<String, Object> response = new HashMap<>();
             response.put("id", id);
+            response.put("message", "map created successfully");
             return new ResponseEntity<>(response, HttpStatus.CREATED);
-        } catch (BadAttributException e) {
+        } catch (BadAttributeException e) {
             System.out.println(e);
             Map<String, Object> response = new HashMap<>();
             response.put("status", HttpStatus.BAD_REQUEST.value());
@@ -88,14 +90,27 @@ public class MapsController {
     @DeleteMapping("/{id}")
     public ResponseEntity<Object> deleteMap(@PathVariable("id") int id) {
         try {
+            if (id < 0) {
+                throw new BadPathVariableException("Pathvariable cannot be negative");
+            }
             Map<String, Integer> rowsAffected = service.remove(id);
             Map<String, Object> response = new HashMap<>();
-            response.put("id", rowsAffected.get("map_id"));
+            response.put("id", id);
             response.put("maps_deleted", rowsAffected.get("maps"));
             response.put("zombies_deleted", rowsAffected.get("zombies"));
             return new ResponseEntity<>(response, HttpStatus.OK);
+        } catch (BadPathVariableException e) {
+            System.out.println(e);
+            Map<String, Object> response = new HashMap<>();
+            response.put("status", HttpStatus.BAD_REQUEST.value());
+            response.put("message", e.getMessage());
+            return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
         } catch (Exception e) {
-            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+            System.out.println(e);
+            Map<String, Object> response = new HashMap<>();
+            response.put("status", HttpStatus.INTERNAL_SERVER_ERROR.value());
+            response.put("message", "internal server error");
+            return new ResponseEntity<>(response, HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 }
