+++ src/main/java/com/epf/api/controller/ZombiesController.java
@@ -5,7 +5,9 @@ import java.util.List;
 import java.util.Map;
 
 import com.epf.api.dto.ZombiesDTO;
+import com.epf.api.exception.BadPathVariableException;
 import com.epf.api.mapper.ZombiesDTOMapper;
+import com.epf.core.exception.BadAttributeException;
 import com.epf.core.model.Zombies;
 import com.epf.core.service.ZombiesService;
 import com.epf.persistance.exception.EmptyDataException;
@@ -35,13 +37,23 @@ public class ZombiesController {
     }
 
     @GetMapping()
-    public ResponseEntity<List<ZombiesDTO>> getAllPlants() {
+    public ResponseEntity<Object> getAllPlants() {
         try {
             List<Zombies> models = service.findAll();
             List<ZombiesDTO> dtos = dtoMapper.mapListModelsToListDTOs(models);
             return ResponseEntity.ok(dtos);
         } catch (EmptyDataException e) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+            System.out.println(e);
+            Map<String, Object> response = new HashMap<>();
+            response.put("status", HttpStatus.NOT_FOUND.value());
+            response.put("message", "no zombie found");
+            return new ResponseEntity<>(response, HttpStatus.NOT_FOUND);
+        } catch (Exception e) {
+            System.out.println(e);
+            Map<String, Object> response = new HashMap<>();
+            response.put("status", HttpStatus.INTERNAL_SERVER_ERROR.value());
+            response.put("message", "internal server error");
+            return new ResponseEntity<>(response, HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 
@@ -58,22 +70,46 @@ public class ZombiesController {
             int id = service.create(model);
             Map<String, Object> response = new HashMap<>();
             response.put("id", id);
+            response.put("message", "zombie created successfully");
             return new ResponseEntity<>(response, HttpStatus.CREATED);
+        } catch (BadAttributeException e) {
+            System.out.println(e);
+            Map<String, Object> response = new HashMap<>();
+            response.put("status", HttpStatus.BAD_REQUEST.value());
+            response.put("message", e.getMessage());
+            return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
         } catch (Exception e) {
-            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+            System.out.println(e);
+            Map<String, Object> response = new HashMap<>();
+            response.put("status", HttpStatus.INTERNAL_SERVER_ERROR.value());
+            response.put("message", "internal server error");
+            return new ResponseEntity<>(response, HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Map<String, Object>> deletePlant(@PathVariable("id") int id) {
         try {
+            if (id < 0) {
+                throw new BadPathVariableException("Pathvariable cannot be negative");
+            }
             int rowsAffected = service.remove(id);
             Map<String, Object> response = new HashMap<>();
             response.put("id", id);
-            response.put("deleted", rowsAffected);
+            response.put("message", rowsAffected + "zombie deleted");
             return new ResponseEntity<>(response, HttpStatus.OK);
+        } catch (BadPathVariableException e) {
+            System.out.println(e);
+            Map<String, Object> response = new HashMap<>();
+            response.put("status", HttpStatus.BAD_REQUEST.value());
+            response.put("message", e.getMessage());
+            return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
         } catch (Exception e) {
-            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+            System.out.println(e);
+            Map<String, Object> response = new HashMap<>();
+            response.put("status", HttpStatus.INTERNAL_SERVER_ERROR.value());
+            response.put("message", "internal server error");
+            return new ResponseEntity<>(response, HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 }
